new file:   leetcode/Easy/cpp/3643. Flip Square Submatrix Vertically.cpp 	new file:   leetcode/Easy/csharp/3643. Flip Square Submatrix Vertically.cs 	new file:   leetcode/Easy/java/3643. Flip Square Submatrix Vertically.java 	new file:   leetcode/Easy/javascript/3643. Flip Square Submatrix Vertically.js 	new file:   leetcode/Easy/python/3643. Flip Square Submatrix Vertically.py 	modified:   online_webinars/10.03.2026/lab_opencl/crypto/aes/aes.c 	modified:   online_webinars/10.03.2026/lab_opencl/crypto/chacha20/chacha20.c 	modified:   online_webinars/10.03.2026/lab_opencl/tests/test_aes.c 	modified:   online_webinars/10.03.2026/lab_opencl/todo.md

+++ online_webinars/10.03.2026/lab_opencl/crypto/aes/aes.c
@@ -127,10 +127,10 @@ int aes256_init(aes256_ctx_t* ctx, const uint8_t key[AES_256_KEY_LEN]) {
 static void add_round_key(uint8_t state[16], const uint32_t* rk) {
     for (int i = 0; i < 4; i++) {
         uint32_t w = rk[i];
-        state[i*4 + 0] ^= (w >> 24) & 0xFF;
-        state[i*4 + 1] ^= (w >> 16) & 0xFF;
-        state[i*4 + 2] ^= (w >> 8) & 0xFF;
-        state[i*4 + 3] ^= w & 0xFF;
+        state[i*4 + 0] ^= (uint8_t)((w >> 24) & 0xFF);
+        state[i*4 + 1] ^= (uint8_t)((w >> 16) & 0xFF);
+        state[i*4 + 2] ^= (uint8_t)((w >> 8) & 0xFF);
+        state[i*4 + 3] ^= (uint8_t)(w & 0xFF);
     }
 }
